feat: add credentials, changed data type dev model

+++ server/models/models.js
@@ -21,8 +21,8 @@ const BasketDevice = sequelize.define('basket_device', {
 const Device = sequelize.define('device', {
     id:{type: DataTypes.INTEGER, primaryKey:true, autoIncrement:true},
     name:{type: DataTypes.STRING, unique:true, allowNull:false},
-    price:{type: DataTypes.INTEGER, allowNull:false},
-    rating:{type: DataTypes.INTEGER, defaultValue:0},
+    price:{type: DataTypes.FLOAT, allowNull:false},
+    rating:{type: DataTypes.FLOAT, defaultValue:0},
     img:{type: DataTypes.STRING, allowNull:false},
 })
